requirements.txt
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
tensorflow==2.6.0
opencv-python==4.5.3.56
google-cloud-vision==2.4.2